requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
reportlab==4.0.7
openai==1.12.0
python-dotenv==1.0.0
Werkzeug==3.0.1
